requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
discord==2.3.2
discord-webhook==1.3.0
discord.py==2.3.2
frozenlist==1.4.0
google-api-core==2.12.0
google-api-python-client==2.104.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.61.0
httplib2==0.22.0
idna==3.10
iniconfig==2.0.0
multidict==6.0.4
oauthlib==3.2.2
packaging==23.2
pluggy==1.3.0
propcache==0.2.0
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
pytest==7.4.3
pytest-mock==3.12.0
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.32.3
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tabulate==0.9.0
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.2.3
websockets==11.0.3
yarl==1.17.2
